requirements.txt
pandas
requests
tqdm
python-dotenv
flake8